Moved `TrimLeadingWhitespace` from `internal` to `commands`
