+++ commands/leading_whitespace_test.go
@@ -1,4 +1,4 @@
-package internal
+package commands
 
 import (
 	"testing"
