+++ internal/acceptance/acceptance.go
@@ -8,12 +8,9 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/databrickslabs/terraform-provider-databricks/commands"
 	"github.com/databrickslabs/terraform-provider-databricks/common"
-	"github.com/databrickslabs/terraform-provider-databricks/internal"
 	"github.com/databrickslabs/terraform-provider-databricks/internal/compute"
-
-	"github.com/databrickslabs/terraform-provider-databricks/qa"
-
 	"github.com/databrickslabs/terraform-provider-databricks/provider"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -75,7 +72,7 @@ func Test(t *testing.T, steps []Step, otherVars ...map[string]string) {
 	stepConfig := ""
 	for _, s := range steps {
 		if s.Template != "" {
-			stepConfig = qa.EnvironmentTemplate(t, s.Template, vars)
+			stepConfig = EnvironmentTemplate(t, s.Template, vars)
 		}
 		ts = append(ts, resource.TestStep{
 			Config:                    stepConfig,
@@ -152,7 +149,7 @@ func AccTest(t *testing.T, tc resource.TestCase) {
 			if s.Config != "" {
 				t.Logf("Test %s (%s) step %d config is:\n%s",
 					t.Name(), cloudEnv, i,
-					internal.TrimLeadingWhitespace(s.Config))
+					commands.TrimLeadingWhitespace(s.Config))
 			}
 		}
 	}
