+++ qa/testing_test.go
@@ -3,7 +3,6 @@ package qa
 import (
 	"context"
 	"fmt"
-	"os"
 	"testing"
 
 	"github.com/databrickslabs/terraform-provider-databricks/common"
@@ -26,38 +25,6 @@ func TestRandomLongName(t *testing.T) {
 func TestRandomName(t *testing.T) {
 	n := RandomName("x", "y")
 	assert.Equal(t, 14, len(n))
-}
-
-func TestEnvironmentTemplate(t *testing.T) {
-	defer common.CleanupEnvironment()
-	err := os.Setenv("USER", RandomName())
-	assert.NoError(t, err)
-
-	res := EnvironmentTemplate(t, `
-	resource "user" "me" {
-		name  = "{env.USER}"
-		email = "{env.USER}+{var.RANDOM}@example.com"
-	}`)
-	assert.Equal(t, os.Getenv("USER"), FirstKeyValue(t, res, "name"))
-}
-
-func TestEnvironmentTemplate_other_vars(t *testing.T) {
-	otherVar := map[string]string{"TEST": "value"}
-	res := EnvironmentTemplate(t, `
-	resource "user" "me" {
-		name  = "{var.TEST}"
-	}`, otherVar)
-	assert.Equal(t, otherVar["TEST"], FirstKeyValue(t, res, "name"))
-}
-
-func TestEnvironmentTemplate_unset_env(t *testing.T) {
-	res, err := environmentTemplate(t, `
-	resource "user" "me" {
-		name  = "{env.USER}"
-		email = "{env.USER}+{var.RANDOM}@example.com"
-	}`)
-	assert.Equal(t, "", res)
-	assert.Errorf(t, err, fmt.Sprintf("please set %d variables and restart.", 2))
 }
 
 func TestResourceFixture(t *testing.T) {
