+++ scim/acceptance/resource_user_test.go
@@ -44,7 +44,7 @@ func TestAccUserResource(t *testing.T) {
 	if _, ok := os.LookupEnv("CLOUD_ENV"); !ok {
 		t.Skip("Acceptance tests skipped unless env 'CLOUD_ENV' is set")
 	}
-	config := qa.EnvironmentTemplate(t, `
+	config := acceptance.EnvironmentTemplate(t, `
 	resource "databricks_user" "first" {
 		user_name = "eerste+{var.RANDOM}@example.com"
 		display_name = "Eerste {var.RANDOM}"
