+++ secrets/acceptance/secret_test.go
@@ -19,7 +19,7 @@ func TestAccSecretResource(t *testing.T) {
 	if _, ok := os.LookupEnv("CLOUD_ENV"); !ok {
 		t.Skip("Acceptance tests skipped unless env 'CLOUD_ENV' is set")
 	}
-	config := qa.EnvironmentTemplate(t, `
+	config := acceptance.EnvironmentTemplate(t, `
 	resource "databricks_secret_scope" "this" {
 		name = "tf-scope-{var.RANDOM}"
 	}
