+++ storage/acceptance/adls_gen1_test.go
@@ -22,7 +22,7 @@ func TestAzureAccAdlsGen1Mount_correctly_mounts(t *testing.T) {
 	acceptance.AccTest(t, resource.TestCase{
 		Steps: []resource.TestStep{
 			{
-				Config: qa.EnvironmentTemplate(t, `
+				Config: acceptance.EnvironmentTemplate(t, `
 				resource "databricks_secret_scope" "terraform" {
 					name                     = "terraform-{var.RANDOM}"
 					initial_manage_principal = "users"
